Moved enrollment specific tests to their own package
