+++ psm-app/integration-tests/src/test/java/gov/medicaid/features/enrollment/steps/DataCoverageStepDefinitions.java
@@ -1,9 +1,8 @@
-package gov.medicaid.features.general.steps;
+package gov.medicaid.features.enrollment.steps;
 
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
-import gov.medicaid.features.enrollment.EnrollmentSteps;
-import gov.medicaid.features.general.ui.EnrollmentPage;
+import gov.medicaid.features.enrollment.ui.EnrollmentPage;
 import net.thucydides.core.annotations.Steps;
